Jim's edits on session processing
I went through the process of doing a session using an example from 2019.  I found that the files and directories would be easier to manage if group in a chunk.  I also added some checks to see if the files exist. I am also thinking we need to standardize the paths and folders.  Other additions needed is more explanations of the various checks so that someone not as familiar with the data can understand the checks.